+++ docs/examples/pic/launch_by_subscription/datasources.tf
@@ -1,0 +1,31 @@
+# Gets a list of Availability Domains
+data "oci_identity_availability_domains" "ADs" {
+  compartment_id = "${var.tenancy_ocid}"
+}
+
+# Gets the partner image listing
+data "oci_core_app_catalog_listings" "test_app_catalog_listings" {
+  filter {
+    name   = "publisher_name"
+    values = ["Oracle CCE Image Management Pipeline"]
+  }
+}
+
+data "oci_core_app_catalog_listing_resource_versions" "test_app_catalog_listing_resource_versions" {
+  #Required
+  listing_id = "${lookup(data.oci_core_app_catalog_listings.test_app_catalog_listings.app_catalog_listings[0],"listing_id")}"
+}
+
+# Gets the partner image
+data "oci_core_app_catalog_subscriptions" "test_app_catalog_subscriptions" {
+  #Required
+  compartment_id = "${var.compartment_ocid}"
+
+  #Optional
+  listing_id = "${lookup(data.oci_core_app_catalog_listing_resource_versions.test_app_catalog_listing_resource_versions.app_catalog_listing_resource_versions[0],"listing_id")}"
+
+  filter {
+    name   = "listing_resource_version"
+    values = ["${lookup(data.oci_core_app_catalog_listing_resource_versions.test_app_catalog_listing_resource_versions.app_catalog_listing_resource_versions[0],"listing_resource_version")}"]
+  }
+}