+++ docs/examples/pic/launch_by_subscription/provider.tf
@@ -1,0 +1,10 @@
+/*
+ * This example file shows how to configure the oci provider to target the a single region.
+ */
+provider "oci" {
+  region           = "${var.region}"
+  tenancy_ocid     = "${var.tenancy_ocid}"
+  user_ocid        = "${var.user_ocid}"
+  fingerprint      = "${var.fingerprint}"
+  private_key_path = "${var.private_key_path}"
+}